% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{GetCIDWithName}
\alias{GetCIDWithName}
\title{GetCIDWithName
Retrive pubchem compound id with chemical name input}
\usage{
GetCIDWithName(chemName, waitTime = 0)
}
\arguments{
\item{chemName}{chemical name}

\item{waitTime}{wait time for each run}
}
\value{
chemName in string with cid as
}
\description{
GetCIDWithName
Retrive pubchem compound id with chemical name input
}
\examples{
GetCIDWithName(chemName = "amoxicillin")

}
