% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{GetNameUrlPugRest}
\alias{GetNameUrlPugRest}
\title{GetNameUrlPugRest
Get the url for Pug Rest APIs with checmial name input}
\usage{
GetNameUrlPugRest(
  chemName,
  searcBasehUrl = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/name/",
  type = "/JSON"
)
}
\arguments{
\item{chemName}{name of the chemical}

\item{searcBasehUrl}{basic search url}

\item{type}{type of the return file}
}
\value{
searchUrl in string
}
\description{
GetNameUrlPugRest
Get the url for Pug Rest APIs with checmial name input
}
\examples{
GetNameUrlPugRest(chemName = "amoxicillin")

}
