% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{GetPubchemCId}
\alias{GetPubchemCId}
\title{GetPubchemCId
Retrive pubchem compound id with chemical name input}
\usage{
GetPubchemCId(chemName, waitTime = 0)
}
\arguments{
\item{chemName}{chemical name}

\item{waitTime}{wait time for each run}
}
\value{
cid in string
}
\description{
GetPubchemCId
Retrive pubchem compound id with chemical name input
}
\examples{
GetPubchemCId(chemName = "amoxicillin")

}
