% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{GetPubRestDescriptionJsonWithName}
\alias{GetPubRestDescriptionJsonWithName}
\title{GetPubRestDescriptionJsonWithName}
\usage{
GetPubRestDescriptionJsonWithName(chemName, waitTime = 0)
}
\arguments{
\item{chemName}{chemcial name}

\item{waitTime}{wait time for each round}
}
\value{
result in json
}
\description{
GetPubRestDescriptionJsonWithName
}
\examples{
GetPubRestDescriptionJsonWithName(chemName = "amoxicillin")

}
