% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{GetPubRestJson}
\alias{GetPubRestJson}
\title{GetPubRestJson}
\usage{
GetPubRestJson(chemName, waitTime = 0)
}
\arguments{
\item{chemName}{chemcial name}

\item{waitTime}{wait time for each round}
}
\value{
result in json
}
\description{
GetPubRestJson
}
\examples{
GetPubRestJson(chemName = "amoxicillin")

}
