% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{GetPugRestUrl}
\alias{GetPugRestUrl}
\title{GetPugRestUrl
Get the url for Pug Rest APIs. For more info about the Pug Rest API: https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest}
\usage{
GetPugRestUrl(
  searchTerm,
  searchDomain = "name",
  returnType = "json",
  returnContent = "description",
  searcBasehUrl = "https://pubchem.ncbi.nlm.nih.gov/rest/pug/compound/"
)
}
\arguments{
\item{searchTerm}{search term}

\item{searchDomain}{domain searched. e.g. name, cid, inchikey}

\item{returnType}{type of the return file. e.g. XML, json}

\item{returnContent}{content to return. e.g. "", "record", "cid", "description", "synonyms",}

\item{searcBasehUrl}{basic search url}
}
\value{
searchUrl in string
}
\description{
GetPugRestUrl
Get the url for Pug Rest APIs. For more info about the Pug Rest API: https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest
}
\examples{
GetPugRestUrl(searchTerm = "amoxicillin")

}
