% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{GetSynonyms}
\alias{GetSynonyms}
\title{GetSynonyms
Retrive all synonyms with input compound id}
\usage{
GetSynonyms(cid, apiKey = "", email = "")
}
\arguments{
\item{cid}{compond id of pubchem database}

\item{apiKey}{optional parameter. ApiKey of the user.}

\item{email}{optional parameter. Email of the user.}
}
\value{
list of synonyms in strings
}
\description{
GetSynonyms
Retrive all synonyms with input compound id
}
\examples{
GetSynonyms(cid = "33613")

}
