% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubchem.R
\name{GetSynonymsWithCID}
\alias{GetSynonymsWithCID}
\title{GetSynonymsWithCID
Retrive pubchem compound id with chemical name input}
\usage{
GetSynonymsWithCID(cid, waitTime = 0)
}
\arguments{
\item{cid}{pubchem compound id}

\item{waitTime}{wait time for each run}
}
\value{
chemName in string with cid as
}
\description{
GetSynonymsWithCID
Retrive pubchem compound id with chemical name input
}
\examples{
GetSynonymsWithCID(cid = "33613")

}
